Restore seam compatibility and break cycles
