+++ tach.toml
@@ -41,6 +41,7 @@ depends_on = [
     "mindroom.media_fallback",
     "mindroom.media_inputs",
     "mindroom.memory",
+    "mindroom.model_loading",
     "mindroom.timing",
     "mindroom.tool_system.events",
 ]
@@ -48,7 +49,6 @@ depends_on = [
 [[modules]]
 path = "mindroom.agents"
 depends_on = [
-    "mindroom.ai",
     "mindroom.agent_prompts",
     "mindroom.constants",
     "mindroom.credentials",
@@ -57,6 +57,7 @@ depends_on = [
     "mindroom.hooks",
     "mindroom.logging_config",
     "mindroom.matrix.identity",
+    "mindroom.model_loading",
     "mindroom.runtime_resolution",
     "mindroom.timing",
     "mindroom.tool_system.catalog",
@@ -74,6 +75,18 @@ depends_on = [
 path = "mindroom.agent_prompts"
 depends_on = []
 visibility = ["mindroom.agents"]
+
+[[modules]]
+path = "mindroom.model_loading"
+depends_on = [
+    "mindroom.constants",
+    "mindroom.credentials",
+    "mindroom.credentials_sync",
+    "mindroom.llm_request_logging",
+    "mindroom.logging_config",
+    "mindroom.vertex_claude_prompt_cache",
+]
+visibility = ["mindroom.ai", "mindroom.agents"]
 
 [[modules]]
 path = "mindroom.agent_policy"
@@ -158,7 +171,7 @@ depends_on = [
 path = "mindroom.api.credentials"
 depends_on = [
     "mindroom.agent_policy",
-    "mindroom.api.main",
+    "mindroom.api.config_lifecycle",
     "mindroom.constants",
     "mindroom.credentials",
     "mindroom.tool_system.worker_routing",
